requirements.txt
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
segmentation-models-pytorch==0.2.1
setuptools==58.0.4
SimpleITK==2.3.1
h5py==3.11.0
torch==2.1.2
monai==1.3.1
mlflow==2.14.1
scikit-learn==1.5.0
lightning==2.3.0
ruamel.yaml==0.18.6
opencv-python-headless==4.10.0.84